rename file containing canari variable adapter
